Update 2.prerequisites.md
Respect to the maintainers. 
I got stumped and looked around for the solution. The documentation indicates to use myapp.dev.test as an alias in the local hosts file, but in docker-compose.dev.yml it has laravel.dev.test as the alias the software actually uses. This can cause the user to be stumped bad right out of the gate. There may be a better or different way to fix this, but there it is.